feat: creation d'une commande multiboot et ajout gestion arithematique sur 64 bits, amelioration de printf

+++ include/types.h
@@ -25,7 +25,7 @@ typedef unsigned long long	UQWORD;
 typedef unsigned char	u8;
 typedef unsigned short	u16;
 typedef unsigned int	u32;
-typedef unsigned long int	u64;
+typedef unsigned long long int	u64;
 typedef char	s8;
 typedef short	s16;
 typedef int	s32;
