+++ include/video.h
@@ -16,6 +16,6 @@ void putchar(u8 thechar);
 u32 print(u8* string);
 u32 printf (const u8 *string, ...);
 void changevc(u8 vc);
-u8* itoa(u32 num, u8* str, u8 base, u32 dim, u8 achar);
+u8* itoa(u64 num, u8* str, u8 base, u64 dim, u8 achar);
 u8* sitoa(int num, u8* str, u32 dim);
